style: 스타일 코드 정리
